Remove logout cleanup call on immediate login failure path [VPNAPPL-1072]
Previously, if a user did not have an active internet connection when
starting up ProtonVPN (i.e., if the computer had just started and was
not yet connected to the internet), the app would clear the user's
credentials from the keychain and reset all local settings. This is
obviously undesirable. We should remove this login cleanup to avoid any
surprises for our users.