Don't report VPN IPs to API [VPNAPPL-970]
Fixes a couple of bugs that were causing VPN IPs to be used for the
server locations API. First was the fact that we did not set the
appStateManager at all on iOS (this dependency was removed), second was
the presence of a race condition that caused the
`NavigationService.launched()` method to race with the
`VPNManager.prepareManagers()` method, resulting in the app thinking the
VPN was disconnected on launch. Finally, though it has no functional
change, the `VPNManager.determineActiveVpnProtocol` method signature was
changed to make the code more self-documenting.