Tweak sysex installation flow [VPNAPPL-1074]
Change version number to 1234 so sysex tour actually closes on
development machines. Fix window closing logic so window close callbacks
actually work. With this fixed, add a callback that sets the profile
back to Ike if the sysex flow is cancelled. While waiting for the tour
to complete, put up a spinner next to the profile selection list. This
is all to allow the profile view to also switch back to Ike if sysex
installation fails, which will be coming in the following commit.