Warn T2 Mac users about unstable VPN configs [VPNAPPL-837]
Pop a dialog on T2 Macs warning that this NetworkExtension + KillSwitch
are known to cause instability (i.e., kernel panics) on their system,
and offer to connect either with IKEv2 or without KillSwitch. Also allow
the user to permanently dismiss these dialogs and always connect
regardless.

(Also fix the "here" hyperlink in the System Extension UI flow to work
in other localizations)