Add sync queues around LocalAgent state changes [VPNAPPL-1034]
The app does not protect against multithreaded access to the LocalAgent
implementation. Therefore it is possible for multiple threads to call
both connect and disconnect on the LocalAgentImplementation
simultaneously, causing the LocalAgent to be torn down while it is
trying to instantiate a connection at the same time, which can lead to
crashes and other unexpected behavior.