Check free space before logging on older devices [VPNAPPL-1255]
The older API calls in Foundation don't provide for catching errors, and
instead raise NSExceptions that aren't catchable from Swift. This means
that attempts to log to the filesystem when it's full, or when the file
handle is otherwise invalid, will crash the system.

This change checks for available free space before each log call on
older devices. It also logs a sentry point on newer devices if a file
error is encountered (and on older devices if there's no free space
left), so we can be made aware of any unknown bugs in our logging code.