Prevent disconnect on cert refresh error in bg [VPNAPPL-1278]
AppDelegate's applicationDidEnterBackground gets called on iOS when the
app enters the background, and applicationDidBecomeActive when same
enters the foreground. If iOS is "pre-warming" our app, it will run app
code in the background (i.e., after the "background" delegate call)
without calling the "active" delegate call.

This change prevents spurious disconnects in the background by
preventing LocalAgent from disconnecting the VPN if a certificate error
is encountered, which can happen due to permissions errors if we're
running in the background.