Define new provider message interface for extensions
This stays compatible with the previous messaging scheme of sending a
single byte at the start of the message to identify it. This is
necessary to allow for slightly more complicated message sending. Handle
these messages in the network extension, holding off on starting the
extension certificate refresh manager until the app has sent a checkin
message, and given a selector to start the API session.

This code stops short of using these new credentials for the refresh
manager, that will come in a later commit.