Display rate-limiting cert refresh alert to user
If the user attempts to change features too many times and this results
in an error 429 from the API, plumb this error condition back from the
extension and display an alert to the user. If the API gives a
retry-after header, display the time contained in this header to the
user as well.