Change link back to "T2 Security Chip"
The "Learn more..." text doesn't fit properly inside the scroll view.
Change the link text back to something that fits for now - other
languages may still look weird if they're longer, but at least the
text scrolls so it isn't completely obscured. The best way forward
would be to consolidate the `WarningPopup` and `Popup` nibs, but that
is non-trivial work.

Critically, the same hack used to make the `Popup` nib resize properly
does not work for `WarningPopup` because of the presence of
`PVPNTextLink` (see other comment fixup changed in this commit):

```
// HACK: Because the text view is inside a scroll view, it doesn't
// resize correctly. To address this, the text view is aligned to
// the text field, which forces the resizing of the dialog.
```