Better cookie handling in ExtensionAPIService
Instead of just sending the session cookie when forking a new session,
handle `Set-Cookie` headers in server responses, and send the session
cookie every time by adding it to a cookie storage object. If the
session cookie should change in the extension (via another `Set-Cookie`
directive), the extension will now pay attention to this header and
update the value.