Use app's API session in NE if creds missing
If the app has not yet checked in with the network extension, and there
are no credentials to be found in the keychain, have the network
extension use the app's API session in the keychain. If the user
launches the app again and the app tries to refresh the certificate, the
app should notice the expired session and fork a new one.