Add VPNManager tests [VPNAPPL-1330]
Tests existing behavior of the VPNManager. Adds an obvious fix where the
WireGuard VPN profile was not being removed from preferences. What still
needs fixing is the strange "hasConnected" behavior with IKEv2, and the
incorrect usage of the `configuration.includeAllNetworks` flag in
VpnManager being used to remove OpenVPN and WireGuard configurations,
instead of keying off of the `includeAllNetworks` setting in the
configuration objects themselves.