Use preferred language as default for country names [VPNAPPL-1313]
A previous "fix" used the current locale (i.e., the "Region" setting in
sysprefs) for country names. This was used to fix a bug where the code
determined the "standard language" code for the preferred language, and
then got the country code, which was incorrect for regionally differing
names of countries. Now, the code uses the preferred regional language
variant for the country name, falls back to the "standard" language name
for the country, and finally if nothing else is found, falls back to the
current locale. Test cases are added to verify this behavior.