Add tests for AppSessionRefreshTimer [VPNAPPL-1354]
Fix a bug where we aren't cancelling the timer that refreshes user data.
The cancel call is made on logout, so this is causing the app to
needlessly attempt to hit the API without valid credentials, for the
purpose of updating the user's vpn credentials.

The tests make sure that server load updates are performed correctly,
and that the requests involving passing the user's IP are correctly
masking the end portion of the user's IP address.