Rewrite SystemExtensionManager [VPNAPPL-1385]
SystemExtensionManager now does the same job as
SystemExtensionStateCheck. Instead of tracking state by saving it on
each request individually, have the request call a `stateChangeCallback`
which allows the SystemExtensionManager to keep track of each request's
state local to the `checkAndInstallAllIfNeeded` call.

Instead of comparing versions using an (unreliable) XPC endpoint before
deciding whether or not to call `checkAndInstallAllIfNeeded`, simply
always call into sysextd to install the extension and have it check
existing installation state itself. Sysextd will then query our
application with the current installed version (if any) and ask if we
want to replace the extension. We will proceed with replacement if we
notice that the version number of the current application is higher than
the version number of the installed extension.

The call to `checkAndInstallAllIfNeeded` will be invoked on app startup
if the user's preferences have a protocol enabled that requires system
extensions, if the user attempts to add a profile with one of these
protocols, or if the user attempts to change their preferences to use
one of these protocols as the default.