UI improvements for sysext flow [VPNAPPL-1385]
Take into account that the "System Extension Blocked" dialog may not
appear if the user has attempted to enable system extensions multiple
times by giving a more helpful guide to the user for how to find
Security preferences on their machine.

Make the ProgressIndicator stop spinning if the user closes the tour
when attempting to create a profile or change their default connection
settings, and revert the preferred protocol back to what it was before
setting it to something requiring a system extension.

Remove the button from the SystemExtensionsTour which says "Begin
Configuration", since now the configuration is automatically started for
the user when the tour is shown.