Permanently allocate SystemExtensionManager [VPNAPPL-1385]
Otherwise system extension requests will get deallocated as soon as the
manager object goes out of scope. Also only call the system extension
manager update on startup if the user has a profile installed that
requires a system extension, or if the settings indicate that the
default protocol requires a system extension.