Log endpoint line in wg config [VPNAPPL-1447]
It looks like the `wireguard-go` backend is improperly parsing the IP
address in the `Endpoint` config line of the wireguard config, but only
intermittently. Log this element of the configuration to determine if
anything strange is happening to it, in case we see it again in the
field a second time.