refactor(profiles): improve code readability
Dealing with indexes of selected items everywhere in the viewmodel got
completely unworkable after a certain point, so I decided to make the
profile view model similar to how the iOS view model worked.

In the process, fix a bug with profile colors not getting populated in
the UI correctly, along with several bugs from the first protocol
override implementation attempt.

Jira-Id: VPNAPPL-1520