fix(services): ensure tour cancellation is called
This change prevents windows from losing references to their delegates
(window controllers) when closeActiveWindows(except:) is invoked, if
they are one of the window types that should remain open.

In particular, this causes the completion handler of the sysex tour
to be called for the tour window that is opened before the user is
logged in and the session changes.