fix(persistence): cache logicals outside defaults
Store server list in a file in the caches directory, instead of user
defaults.

After logicals reached ~2500 servers, the size of the serialized server
list exceeded 4MiB, causing any data being saved to user defaults to
exceed the maximum allowed value. This caused the system to transition
to standard mode, meaning the iOS app and extension would not be able
to read and write to the shared user defaults bucket, breaking some
functionality such as profiles and reverting settings.

Jira-Id: VPNAPPL-1676