fix(certrefresh): certificate feature decoding
Stop erasing certificate features when requesting a new certificate
on MacOS. This prevents decoding failures whenever we attempt to load
these features from storage.

Since the MacOS app, and therefore also LA, is guaranteed to be active
when the NE is launched, there is no need to refresh the certificate
when current features no longer match the stored certificate's features.