build: make vpncore a Swift package, LegacyCommon
Move all files to the new location, `libraries/LegacyCommon`. Rename all
sites using the name `vpncore`. Move all of the resources from the
old framework location (in the `Localization` folder) to the Resources
folder, though these will be moved to the existing Strings package
shortly.

While moving this, the mocks that were previously used by the app unit
test targets and lived in vpncore have now been broken out into a
separate TestSupport module, along with all of the mocks used in the old
vpncore unit tests themselves. Because of this, a lot of them have been
modified to have `public` properties and methods.