feature: support for vpn business gateways
Allow truncating of excessively long server names. Also add
business branding and info icons explaining what gateways are. Use the
new feature authorizer code to make checking for plan eligibility a
little bit more concise. This also required a good deal of changes to
all of the `*PropertyProvider` objects.

The gateways section of the server list is implemented using the new
ServerGroup type, which is now a proper struct that allows for dividing
the list more easily into arbitrary sections.

Jira-Id: VPNAPPL-1841