fix: move mocks back into LegacyCommon
For some reason, Xcode couldn't correctly resolve library dependencies
when building the macOS unit tests and linking LegacyCommonTestSupport.
The iOS unit tests and LegacyCommon unit tests worked fine, but for some
reason the macOS unit test build was unable to find transitive
dependencies of Swift packages during link time, leading to the linker
complaining about not being able to find `SwiftyBeaver` or other
transitive dependencies of the LegacyCommon package.

Move all of the mocks into `LegacyCommon` for now, and keep them behind
an `#if DEBUG` compiler check. This is the best we can do for now,
unfortunately.