refactor(core): introduce storage dependencies
This commit introduces refactored interfaces to replace Storage.
 - DataStorage: 'big data' storage with a liveValue implementation of
`FileStorage`, a temporary solution for storing servers, to be replaced
by a dedicated persistance layer such as Core Defaults in the future.
 - Storage: a wrapper around `UserDefaults`, a direct successor to the
old Storage implementation.
 - DefaultsProvider: A replacement interface for where the old Storage
implementation's `defaults` property was used to directly expose a
specific instance of UserDefaults.

This refactor is mainly made possible by the ability to separate the
DefaultsProvider interface defined in one of the leaf node packages of
NEHelper, with live implementations living in the respective app
targets, with iOS (and therefore its network and app extensions)
depending on an implementation that specifies a shared container.