test(ios): fix test build issue
Unit tests for iOS were in a strange place, where the test target
relied on the host app, but also specifically compiled a set of source
files from the Application target, on top of test files.

This commit fully moves testing strategy to being based on the host
app and importing its API's through @testable import ProtonVPN
statements. In the future, we should consider moving to the alternate
approach, to being able to run unit tests quickly without building the
application and its dependencies.

In addition, this commit stops us from linking against some of our
libraries that are already linked due to relying on the host app, like
LegacyCommon and Theme. These were causing issues where e.g. alerts
initialised in the app target were failing to be cast in LegacyCommon
and were causing test failures (and potentially preventing the correct
alert from being shown when running in release mode) due to duplicated
definitions of the build objects.