refactor(core): feature authorizer interface
Adds a @Dependency interface for controlling feature enablement for
features based on remote FeatureFlags and AccountPlan, along with two
related intermediate dependencies:
 - FeatureAuthorizer: controlling feature enablement for features based
on remote FeatureFlags and User Tier
 - PlanProvider: A @Dependency alternative to UserTierProvider. Its live
 implementation calls out to the VpnKeychain singleton to retrieve the
user's account plan.
 - FeatureFlagProvider: Provides an interface to get and set remote
feature flags. Its live implementation holds onto an instance of the
PropertiesManager.

This commit also converts the live implementation of VpnKeychain to a
singleton, to prevent the Dependencies library and the legacy
DependencyContainer from creating more than one shared instance.

Rename FeatureAuthorizer to FeatureAuthorizerProvider, since it does not
directly authorize the user of features, but rather is used to generate
authorizers.

Update the intended manner in which authorizable app features are
defined, providing two alternative protocols to conform to, AppFeature
and ModularAppFeature depending on whether the feature is self-contained
or has sub-features/levels, instead of forcing all features to define
a (possibly Void) SubFeature type.