feature: check for bad LAN configurations on connect
If an interface has an improperly configured subnet range, this can
result in the operating system sending packets over the network outside
of the tunnel.

Double-check before connecting that none of the interfaces have anything
weird defined, and if they do, display a warning to the user encouraging
them to connect using Kill Switch.

Jira-Id: VPNAPPL-1765