[6cdfe11b1] fix(refresh): change refresh criteria
We already check if we're logged in elsewhere during app session
refresh, so don't hit the keychain if we can avoid it. Also dial up the
recent activity threshold for macOS and remove some dead code in server
features.

Jira-Id: VPNAPPL-1799
[e4209aada] fix(p2pcheck): reduce frequency & load of p2p checking

Jira-Id: VPNAPPL-1799
[0ed72c2c6] fix(clientconfig): only fetch when app is recently active

Jira-Id: VPNAPPL-1801
[cb91282b7] fix(serverlist): don't fetch paid servers for free users

Instead of getting the full server list, reduce load on the backend if
the user is a free user by only fetching the free server list.

Jira-Id: VPNAPPL-1801
[e5bad1d02] fix(refresh): reduce frequency of server info fetching

Reduce the frequency of calls that refresh the available partners and
streaming services information.

Jira-Id: VPNAPPL-1803
[f85684186] fix(maintenance_check): add jitter to refresh interval

Jira-Id: VPNAPPL-1800