feature(settings): make some features paid
This commit adds `AppFeaturePropertyProvider` and `ProvidableFeature`
which works together to standardise the definition of app features that
can be turned on and off by feature flags, made paid based on user tier
and plan, and can be stored per user in `Storage`.

VPNAccelerator and LANConnections (aka ExcludeLocalNetworks) have been
defined as such features, ensuring they are now stored per user, and are
gated to only be available to paid users. VPNAccelerator will also now
be turned on by default for all new users, as previously intended.

Jira-Id: VPNAPPL-1852