feature: authenticate widget deeplink clicks
A deeplink handler in the iOS application was directly accepting
requests to connect and disconnect from the VPN. We use these deeplinks
from the widget to let the app handle connection and disconnection
requests from the widget.

The widget will now include three pieces of information when sending a
connection or disconnection request to the application:

- The current unix timestamp,
- A signature of the timestamp,
- The algorithm used for that signature.

The key pair is generated inside the widget, and the widget's public key
is placed in the app's keychain. When the app handles the deep link, it
checks if the timestamp was recent, then verifies the timestamp's
signature using the public key. If the signature is missing or invalid,
the deep link action will fail to run.

Jira-Id: VPNAPPL-1935