fix: plumb TLS_PIN_DISABLE flag into LegacyCommon
Moving vpncore (the framework) to LegacyCommon (the Swift package)
caused it to ignore the TLS_PIN_DISABLE flag because it was no longer
part of the main project. This change plumbs the value through to from
the app layer to the module through an additional argument in the
CoreNetworking init routine.