fix(certauth): recover from certificate mismatch
Certificate refresh operations are susceptible to data races, with one
particular case commonly observed in production: a certificate generated
against one pair of keys being saved to the keychain after these keys
have been replaced by another set due to logout or plan change.

Once this occurs, if LocalAgent is provided a certificate that does not
match our current keys, it will fail to connect resulting in a loss of
connectivity.

This commit mitigates the impact of mismatched keys by validating the
certificate against our keys to verify the public key matches, and
deleting it if necessary.

Jira-Id: VPNAPPL-1793