refactor(storage): prevent excessive warnings
On iOS, we store the set of features that our current VPN certificate
was generated against. On both platforms, we persist information about
our current connection, with one bucket for each protocol (ike, openvpn,
wireguard).

For both of these, it is reasonable to store nil values. Currently,
whenever we attempt to load one of these values and it is nil, we will
output warning logs.

Update set<T: Encodable> to explicitly accept optional values, and
represent nil values in user defaults by saving nil, instead of
encoding the string "null", for a small performance improvement. Also,
when decoding, attempt to decode T? instead of T to silence warnings
due to attempting to decode nil values persisted before this change.

The alternative would be to provide optional and concrete overloads of
the get and set functions.