fix: update ios test target TEST_HOST path
This commit fixes the `Missing Required Module` build errors in the iOS
ProtonVPNTests target.

Xcode 14 has a bug where the TEST_HOST environment variable is
incorrectly defined, causing the value of Bundle_Loader to be invalid
and failing builds due to missing. The path includes
a component consisting of the variable BUNDLE_EXECUTABLE_FOLDER_PATH,
which evaluates to an empty string and the ultimate value of TEST_HOST
ends up invalid with a double slash "...ProtonVPN.app//ProtonVPN"

This affects non-swift frameworks, since it meant the modulemap for
these frameworks could not be found when building the test target.