fix(status): break iOS retain cycle
Status View Model was being retained in two ways:
 - timer was being setup with the self as the target. This causes the
timer to hold a strong reference to the view model until `invalidate`
is called. However, the viewmodel did not do this, since the call to
invalidate the timer was in its deinit method.
 - the view model holds a strong reference to notification tokens, for
which the handlers held a strong reference to the view model.

This meant that each time the user navigated to and away from the
`StatusViewController`, a new `StatusViewModel` was initialised and
held onto forever, while also firing timer events, querying connection
time, and responding to notifications.

This commit breaks those retain cycles by:
 - using the closure based Timer API with a weak reference to the view
model in the handler closure
 - changing the notification handlers to also hold a weak reference to
the view model